CLIP-for-embedding-generation/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements/requirements.txt requirements/requirements.in
#
accelerate==1.2.1
    # via -r requirements/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   pip-tools
    #   uvicorn
fastapi==0.115.6
    # via -r requirements/requirements.in
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via uvicorn
huggingface-hub==0.27.0
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.5
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.1
    # via
    #   accelerate
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via
    #   accelerate
    #   build
    #   huggingface-hub
    #   transformers
pillow==11.1.0
    # via -r requirements/requirements.in
pip-tools==7.4.1
    # via -r requirements/requirements.in
psutil==6.1.1
    # via accelerate
pydantic==2.10.4
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.0
    # via
    #   accelerate
    #   transformers
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.1
    # via torch
tokenizers==0.21.0
    # via transformers
torch==2.5.1
    # via
    #   -r requirements/requirements.in
    #   accelerate
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.47.1
    # via -r requirements/requirements.in
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements/requirements.in
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
